Reorganize RefChecks tasks
 - Split out forward reference checks and cross version (i.e. experimental/deprecated) checks
   into their own miniphases. These have nothing to do with the core tasks of RefChecks.
 - Move RefChecks earlier in the pipeline, so that it is not affected by ElimByName's questionable
   type manipulations.
